+++ src/app/modules/student-answers/dtos/get-all-student-answer.dto.ts
@@ -1,4 +1,0 @@
-import { GetBaseStudentAnswerDto } from './get-base-student-answer.dto';
-
-export interface GetAllStudentAnswerDTO extends GetBaseStudentAnswerDto {
-}