+++ src/app/modules/student-answers/models/student-answer.ts
@@ -1,6 +1,8 @@
 import { GetBaseQuestionDTO } from '../../questions/dtos/get-base-question.dto';
-import { GetAllStudentAnswerDTO } from '../dtos/get-all-student-answer.dto';
-import { Text } from './text';
+import { GetStudentAnswerDetailedDTO } from '../dtos/get-student-answer-detailed.dto';
+import { GetStudentAnswerSummarizedDTO } from '../dtos/get-student-answer-summarized.dto';
+import { SimilarityMatricesDTO } from '../dtos/similarity-matrices.dto';
+import { ProcessingStatus, Text } from './text';
 
 export class StudentAnswer {
 
@@ -13,8 +15,10 @@ export class StudentAnswer {
 	question: GetBaseQuestionDTO;
 	text: Text;
 	student: number;
+	similarityMatrices?: SimilarityMatricesDTO;
+	similarityMatricesStatus?: ProcessingStatus;
 
-	static fromDto(
+	static fromSummarizedDTO(
 		{
 			uuid,
 			student,
@@ -22,7 +26,7 @@ export class StudentAnswer {
 			question,
 			grades,
 			grade
-		}: GetAllStudentAnswerDTO
+		}: GetStudentAnswerSummarizedDTO
 	): StudentAnswer {
 		return new StudentAnswer({
 			uuid,
@@ -30,7 +34,32 @@ export class StudentAnswer {
 			text,
 			grade,
 			grades,
-			student
+			student,
+
+		});
+	}
+
+	static fromDetailedDTO(
+		{
+			uuid,
+			student,
+			text,
+			question,
+			grades,
+			grade,
+			similarityMatrices,
+			similarityMatricesStatus
+		}: GetStudentAnswerDetailedDTO
+	): StudentAnswer {
+		return new StudentAnswer({
+			uuid,
+			question,
+			text: Text.fromDetailedDTO(text),
+			grade,
+			grades,
+			student,
+			similarityMatrices,
+			similarityMatricesStatus
 		});
 	}
 
@@ -41,5 +70,7 @@ export class StudentAnswer {
 		this.question = studentAnswer.question;
 		this.text = studentAnswer.text;
 		this.student = studentAnswer.student;
+		this.similarityMatricesStatus = studentAnswer.similarityMatricesStatus;
+		this.similarityMatrices = studentAnswer.similarityMatrices;
 	}
 }