+++ src/app/modules/student-answers/services/student-answers.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { ApiService } from '../../dataset-manager/services/api.service';
-import { GetAllStudentAnswerDTO } from '../dtos/get-all-student-answer.dto';
-import { GetStudentAnswerDTO } from '../dtos/get-student-answer.dto';
+import { GetStudentAnswerDetailedDTO } from '../dtos/get-student-answer-detailed.dto';
+import { GetStudentAnswerSummarizedDTO } from '../dtos/get-student-answer-summarized.dto';
 import { PatchStudentAnswerDTO } from '../dtos/patch-student-answer.dto';
 
 @Injectable({
@@ -15,18 +15,18 @@ export class StudentAnswersService {
 	) {
 	}
 
-	getAllByQuestionUuid(questionUuid: string | undefined): Observable<GetAllStudentAnswerDTO[]> {
+	getAllByQuestionUuid(questionUuid: string | undefined): Observable<GetStudentAnswerSummarizedDTO[]> {
 		if (!questionUuid) {
 			return of([]);
 		}
-		return this.apiService.get<GetAllStudentAnswerDTO[]>(`/questions/${questionUuid}/student-answers`);
+		return this.apiService.get<GetStudentAnswerSummarizedDTO[]>(`/questions/${questionUuid}/student-answers`);
 	}
 
-	updateById(uuid: string, patch: PatchStudentAnswerDTO): Observable<GetStudentAnswerDTO> {
-		return this.apiService.patch<PatchStudentAnswerDTO, GetStudentAnswerDTO>(`student-answers/${uuid}`, patch);
+	updateById(uuid: string, patch: PatchStudentAnswerDTO): Observable<GetStudentAnswerDetailedDTO> {
+		return this.apiService.patch<PatchStudentAnswerDTO, GetStudentAnswerDetailedDTO>(`student-answers/${uuid}`, patch);
 	}
 
-	getById(uuid: string): Observable<GetStudentAnswerDTO> {
-		return this.apiService.get<GetStudentAnswerDTO>(`student-answers/${uuid}`);
+	getById(uuid: string): Observable<GetStudentAnswerDetailedDTO> {
+		return this.apiService.get<GetStudentAnswerDetailedDTO>(`student-answers/${uuid}`);
 	}
 }