+++ src/app/modules/student-answers/state/student-answers.state.ts
@@ -1,0 +1,14 @@
+import { ResourceStatus } from '../../../state/global.state';
+import { StudentAnswer } from '../models/student-answer';
+
+export interface StudentAnswersState {
+	studentAnswerShown: StudentAnswer | null;
+	studentAnswerShownStatus: ResourceStatus;
+	sendingProcessTextRequest: boolean;
+}
+
+export const studentAnswersInitialState: StudentAnswersState = {
+	studentAnswerShown: null,
+	studentAnswerShownStatus: ResourceStatus.NotLoaded,
+	sendingProcessTextRequest: false
+};